moved atomic trajectory function and refactored code
